rule: retrieve and log all errors of a gojq query
Running the query rule on a JSON event payload sometimes failed because the
query returned multiple results.
It is not clear why this is happening, when the same query was run with gojq on
the same payload locally it succeeded and returned only a single result.

Maybe the iter contains a context timeout/cancelled result?

Collect all results that a query returns log all errors it returns.
If it returns multiple result that aren't errors, log those.