provider: remove unnecessary JSON marshall of github webhook event
It is not necessary to marshal the github event from it's struct representation,
returned from github.ParseWebHook(), to JSON.

The payload returned by github.ParseWebHook() is always in JSON format.
Use this JSON payload as part of the event that is forwarded to the eventloop.

This seems to fix an issue where running the gojq query returned a "cannot
iterate over nil" error.