event: refactor extracting Event information from github events
Instead of checking if the github webhook event is a specific type of the github
package, check if it has the method from that we get a specific information.

Multiple types e.g. have a GetRef() method, checking for the existence of the
method allows to extract the information from all types that have it, instead of
having to add manually every type to the previous type-switch.

Event.Logfields was converted to a function.
It creates the LogFields when it is called. This allows a better abstraction
instead of requiring that every provider implementation has to set the
same logfields based on the Event fields.

Event.PullRequestName was removed, it was not used.