evloop: give up immediately if an action returns a context cancelled error
If action.Run() returns a context canceled error, give up immediately executing
the action and do not log 2x that the action was canceled.

If this happened so far, the error was logged, the loop iteration was skipped,
then the canceled error was received from ctx.Done() and we logged another time
that the context was canceled and gave up.