action: implement "updatebranch" action
The action updates the branch of a github pull-request with it's base branch.
The updatebranch action uses a developer-preview GitHub API feature and might
change in the future.

It requires a github API token that can be specified in the goordinator config
via the "github_api_token" field.

If the branch is already uptodate nothing is done, otherwise the update
operation is scheduled via the Github API.
The action does not poll for the completion of the operation and assumes that it
always succeeds when it is triggered.
The API documentation is not clear if a scheduled update operation can also
fail. If this is the case, we can poll for the completion in the  future.

The update branch action can fail, if the branch was changed while Goordinator
tried to update it.
If this happens, the action will be retried by the event loop.

The github clients return an goorderr.RetryableError if an operation can be
retried. If this happens because the github API rate limit was exceeded, the
error also contains the information when the ratelimit is resetted and should be
retried again.

The event loop was not adapted yet to handle the RetryableError. It retries
actions all error, as before.
It will be changed in a follow-up PR to only rerun RetryableErrors.