httprequest: return RetryableErrors from Runner
In a following PR the event loop will be changed to only retry actions that
returned a goorderr.RetryableError.

Wrap errors that are returned when sending a HTTP-Request fails or returned an
error status code in RetryableErrors.
For now all errors are wrapped and retried, in the future the http response
status codes on which the requests will be retried can be made configurable in
the cfg file.