evloop: only retry action if a RetryableError was returned
The eventloop only retries action when they return a RetryableError or an error
that wraps a RetryableError.
If a RetryableError has a non-zero After field, the action is only re-executed
after this point in time. This is especially useful for delaying retrying github
API operations until the rate-limit is resetted, if they failed because the rate
limit was exceeded.