autoupdater: introduce github pull request branch autoupdater
This commit introduces a new major feature for Goordinator.
The autoupdater keeps Pull-Requests up to date with their base-branch.
Pull-Requests are added to a per base-branch-queue and the first pull-request in
the queue is kept uptodate with it's base-branch.
If merging the base-branch into the PR branch fails, it's check status becomes
negative or the PR became stale, updates for it are suspended.
Updates for it are resumed when the base-branch or the PR branch changed or the
check status of the PR became positive.

Autoupdate is used together with GitHub's auto-merge feature or a comparable
The autoupdater serializes updates per base-branch, to avoid a race between
pull-requests to get updates the fastest and have a successful CI check first.