events: move evloop specific github event processing logic to goordinator pkg
The functionality of the provider/github package is minimized to only receive,
validate and parse the webhook events.
The provider stores the DeliveryID, JSON representation and the parsed webhook
in a struct that is forwarded to an event channel.

The logic to parse the webhook event struct and store information like the
Pull-Request-Number, Branch, etc into a struct that is used for templating
actions is moved to the goordinator package.

This will allow in a following PR to also forward the events from the
provider/github to a branch-autoupdater implementation, which will not use the
additional fields that are parsed and stored in the goordinator.Event struct.
provider/github now also supports to forward the events to multiple channels.
The provider channel is created in the main package, instead of by the
Evloop. This should make it easier to implement a graceful shutdown in the
future.

The event parameter type of the Template() method of the ActionConfig interface
is changed to be of type action.Event.
action.Event is an interface that specified the required getter-methods.
This was done to not introduce a dependency of the action package to the
goordinator package.
The Getter-Methods are prefixed with "Get" because the names without the Get
prefix are already used by public fields, used by the go-templater.