retryer: reduce logging
Reduce the logging that the retryer does.
The caller can log the execution result instead. The caller has more context
available for the log message and can decide when and what should be logged
freely.
The retryer now only logs messages with debug log level for events that the
caller is not aware of.