evloop: fix: ensure scheduleAction() does not modify event.LogFields
In scheduleAction() elements were appended to event.LogFields and the returned
slice was passed to retryer.Run().

If the event.LogFields slice has >0 capacity the underlying array is not copied.
This can cause that a following scheduleAction() call will modify during the
same operation the slice from a previous call.
retryer.Run() then could log fields from a wrong execution.

Append to the value of action.LogFields() instead of to event.LogFields.
action.LogFields() always returns a new slice.