autoupdater: log reason why a pr is suspended and resumed
When goordinator was not running with debug log level mode it was not possible
to understand from the log why updates for PR were suspended and resumed.

Change the logging behavior to always log from the caller, after a
queue.Suspend() and queue.Resume() call the reason why a pr was suspended.