autoupdater: log reason why a pr is enqueued, dequeued and moved
When goordinator was not running with debug log level mode it was not possible
to understand from the log why a pr was enqueued, dequeued or moved to another
base-branch queue.

Change the logging behavior to always log from the caller, after a
queue.Enqueue() and queue.Dequeue() call was done the reason why a pr was
suspended.