autoupdater: fix: negative status check for old commit suspends updates
Github Status check webhook events contain a list of branches that contain the
commit for that a status check is negative.
When the event was received autoupdater suspended updates for pull-requests that
match the given branch.

It might be that the status check notification is for an older commit and not
the HEAD commit of the given PR. If this happened, updates for a PR were
suspended despite the status for the HEAD commit might be positive.

Instead of suspending a pull-request for updates when a negative status webhook
event was received, the autoupdater now runs the update action for all
pull-requests that match the branch, if the pull-request is first in the queue.

This ensures that a PR is only suspended if is uptodate and it's head commit has
a negative status check.

Pull-Requests that are not the first element in the queue will be kept active,
despite a failed status check. Suspending is delayed.
When they become the first element in the queue and the status check is still
negative for the head commit, updating will be suspended.