github: do not log webhook event payload
Logging the github webhook event payloads with debug priority added a lot of
noise.

Do not log them anymore and ensure the github webkook delivery ID is logged in
the first "received a http request" debug log message.
With the delivery ID it is possible to look up the payload on the github
website (if is a valid webhook event from github).