goordinator: reduce unnecessary nil-checks when parsing gh events
The github event Getter() handle safely nil fields and are returning an empty
value when called on a nil-field.

Reduce the unnecessary nil checks, when multiple Get calls are done on the same
field the nil check is still done, to reduce the number of calls.
(won't really make a difference)