github: determine if a PR uptodate via compare API endpoint
The implementation to determine if a PR was uptodate with it's base branch was
not reliable.
In the scenario documented in the commit message of commit 2d9ff1a
PRIsUptodate() returned true despite the PR branch was not uptodate.

Use the compare github API rest endpoint to determine if a PR is uptodate.
When comparing the base branch with the PR branch it returns the number of
commit the pr branch is behind the base branch.
If it is more =>1, the PR is not uptodate.