autoupdater: honor PR review status
The autoupdater was ignoring the review state of pull-requests.
This caused that when:
- all status checks for a PR were successful,
- it was the first element in the queue,
- and review approval was configured as merge requirement in GitHub

that the PR got stuck at the first place, until the stale timeout was reached
and the PR was suspended.
Github was not automerging the PR because the approvals were missing.

When evaluating the status of a PR the approval status is now queried via the
github search API. If the PR is not approved it is suspended.
If a review was submitted and the PR status is approved autoupdating is resumed.
If a dismissed event for a review is received and the PR is the first one in the
queue, it's approval status is reevaluated.
If a negative review (change request) event or a dismissed event for a review is
received, the approval status is checked again if the PR is first in the queue.

This change means that PRs can be enqueued for autoupdating despite their review
status is not approved, without blocking the queue Unnecessarily.