autoupdater: query combined approved status from graphQL instead of REST API
The REST search API sometimes returned old cached result for the approval
status.
After the event that a PR was approved or changes were requested was received,
the search still returned the previous status.
This can cause that PRs are put into the wrong queue.

Via the GraphQL endpoint I encountered recently the same issue but now I can not
reproduce it anymore.
It is not clear if it can also return a cached status or if that was a temporary
issue on GitHub.
If the GraphQL API has the same issue, possible workaround could be:
- get the updatedAt field via graphql for the last changed pull request review
  when querying the approval status and check that it is later then the time the
  last review even was received for the PR,
  if it's not later, query the approval status again,
- add a configuration option to manually configure the number of required review
  per repository, get from the API the number of review and if a PR has X
  approvals and no changes-requested reviews