retryer: fix: action ran in tight loop if retry did not succeed after 15min
When an action was retried unsuccessfully for more then 15min, the action was
run in a tight loop. It could consume an unreasonable amount of CPU and produce
tons of log output.

This happens because the exponential backoff timer used it's default max elapsed
time setting of 15min.
After the 15min the backofftimer returned a Stop (-1 int) constant.
The retryer did not check if Stop or a negative value was returned and set the
timer for the next retry to a negative value, which means instantly.

1.) Disable the max elapsed time in the backoff timer by setting it to 0.
2.) Check in the Retryer if Stop was returned by the back Backoff timer as
  additionally safety check. This should never happen because of 1.