autoupdater: process checkRun events
The autoupdater now also receives and processes checkRun github events,
analogous to status events.

If a negative check run result event is received, a PR is suspended if it is queued.
If a successful check run result event is received, a PR is enqueued if it is
currently dequeuend and has a combined successful status state.

In a follow up PR the autoupdater will also the check result when deciding if a
PR should be resumed or be dequeued when the update action is run.