autoupdater: update stale timeout when receiving events
The autoupdater was using the timeout returned by the CombinedStatus of the
githubclient to update the stale timeout.
The CombinedStatus was queried the stale timeout was set to this value.

In a following commit the CombinedStatus method will be removed.
To make this possible, the stale timeout is now updated when CheckRun or Status
Events are received instead.