autoupdater: consider github check status of pull-requests
The autoupdater was only querying the status check state of a PR to evaluate if
a PR should be kept in the merge-queue or be suspended.
If a PR had a negative check status result, it could get stuck in the first
place of the queue until the stall timeout expired.

The CombinedStatus and ReadyForMergeStatus methods of the GithubClient are
replaced with a single ReadyForMergeStatus() method.
This methods queries GitHub via GraphQL and returns the review and
StatusCheckRollup state.
The StatusCheckRollup state is an overall state combining the check run results
and status states.

The autoupdater calls ReadyForMergeStatus instead of the previous both methods
when running the updatePR action to decide if a PR should be suspended because
it is missing review or it StatusCheckRollup is negative.
ReadyForMergeStatus is also called when checking if updates for a suspended PR
should be resumed.

The StatusCheckRollup results PENDING and EXPECTED are handled as pending
results. ERROR and FAILURE are handled as failures.