rules/templating: add queryescape function
In the rule.action fields of the goordinator config, strings can now be
escaped for url query parameters by calling the queryescape template function.

This can be useful when fields from the event should be used as parameters in a
trigger URL but contain characters that must be escaped in urls.

The queryescape template functions calls the url.QueryEscape function from the
Golang stdlib.