autoupdate/base-branch-change: clarify log message when PR is not queued
When the base branch for a PR is changed that is not queued for update,
"changing base branch failed" was logged.
If the PR is not queued, log the message only with debug instead of info
priority and add the information to the message that the PR is not queued.