autoupdater: honor status check branch protection rules
Change the autoupdater to honor the required status checks configured as
branch protection rules. Updates for PRs are not suspended anymore if a status
check fails, that is **not** a required status check for the base branch.

So far the autoupdater was using the statusCheckRollupState of a PR to determine
if CI checks succeeded. The StatusCheckRollupState does not take the required
status checks configured as branch protection rules into consideration.
If a CI check fails that is not configured as being required, the
StatusCheckRollupState is failed.

To implement the new behavior, the list of of required status checks for the
base branch for the PR is additionally queried via graphql and compared with the
submitted statuses.
The overall CIstatus is pending, if StatusCheckRollupState is pending or if an
required check has not submitted any status yet and no required check failed.
If all required checks succeeded and optional ones failed, the status is
success.