autoupdate: check if PR is first when running update operation
It can happen that the Update operations is run for multiple PRs instead of only
the first one in the active queue.
An example is that multiple update operations are queued in
the go-routine pool. The first run is running for PR #1, the queued one is for PR #2.
If PR #2 gets suspended, the update operation is still queued.
When a PR is suspended the update operation is only canceled if it is
also first in the queue and has been started.

To prevent it, updatePR() now checks if the PR for that it runs is the first one
in the active queue, otherwise it does nothing.
It also now checks if it's context is canceled, in the beginning, to prevent
unnecessary function calls

A better solution would be to change the go routine pool, to be able to remove
operations for the queue via the PR identifier. This would allow to always
remove updatePR operations when a PR gets suspended.