autoupdate: fix: branch update and status check failure race
Sometimes multiple PRs in the autoupdater queue are updated, instead of only the
first in the queue.
The operation to update a PR branch with it's base branch was not blocking.
The GitHub API either returned a 200 status, when the branch update was
triggered and done or 202 to indicate that it is scheduled.
When it was scheduled, the update action continued and the work queue that
synchronizes update operations processed the next update operations.
When goordinator is also used to trigger automatically CI jobs for PRs that were
updated and for that automerge is enabled, and CI jobs report a fail status when
they are aborted because of an update, it also causes unnecessary CI runs.

The race works like this:
- PR#1: PR update with base is triggered, updates gets scheduled by Github
- PR#1: github updates the branch
- PR#1: CI checks fail notification for PR #1 for outdated HEAD commit
  is received, caused by abort of CI jobs for the outdated HEAD
- PR#2 becomes #1 in the queue, because PR#1 was suspended,  update is triggered
  for PR#2, causing start of CI jobs
- PR#1: Is added back to the active queue and resumed, because it's branch changed
- PR#1: CI jobs are started because the branch changed
=> The first and last PRs in the queue have been updated with it's base, for
   both CI jobs are running.

The race is fixed by always waiting in the Update operation until the branch has
been updated. When a scheduled status is returned from GitHub, the branch status
is polled periodically until it is up2date, an error is returned, or the retry
timeout expired.