autoupdate: label first PR in the queue
The first PR in the autoupdate queue now gets a GitHub label applied.
When the first place changes the label is removed and the first PR gets the
label applied.

The label will allow to coordinate actions between autoupdater and the rules, to
e.g. only trigger CI runs when a PR is first in the queue.