autoupdate: do not abort suspend procedure in unlikely error case
If a PR should be in the suspend and active queue because of a bug, continue
with the suspend procedure and overwrite the existing PR in the suspend queue.
This ensures that everything continues to work, instead of not not removing the
PR label and not aborting an update task for it.

The error scenario never happened.