githubclt: fix: UpdateBranch logs debug up2date message with old head
UpdateBranch is retrieving the HEAD sha of the PR and then checking if the PR
branch contains all changes of the base branch by specifying the PR branch name
as reference.
Between retrieving the head sha and comparing the branches, the branch might
have changed which causes that UpdateBranch was logging that the branch is
up2date with an old head commit.

Fix it by passing the PR head SHA instead of the branch name as reference when
comparing the branch with it's base.