autoupdate: always debug log ready for merge status result
Instead of logging the ready for merge status only in
resumeIfPRMergeStatusPositive, always log it with a debug message when it
succeeds.

The review decision, ci_statuses and ci status summary fields are also not added
to the logger anymore in updatePR() and included in all further log messages.
It is unnecessary to include it in multiple log messages.