autoupdate: log warning if retrieved status is for different commit
If the status is fetched for a different commit then the current HEAD commit,
log warning message.
This might happen when the GitHub API returns outdated information, or the PR
branch changed between both API calls.

Goordinator is currently not handling the situation. It can cause that a PR is
suspended wrongly because of the status of a previous commit.