autoupdate: fix: updatePRWithBase wrong changed state
If an update of a PR with it's base branch is scheduled and UpdateBranch is
called again until it returns that the branch is uptodate, updatePRWithBase
returns wrongly that the PR did not change.

This causes that updatePR() continues to run with the ready for
merge status of the previous HEAD commit.
If that one had failed checks it causes the PR to be suspended.

Fix it by returning changed=true if updatePRWithBase scheduled a update or the
branch was updated immediately via the github client.